Огромный рефакторинг parser
