+++ parser/domain/model/processing_result.py
