+++ parser/gateway/gemini/embedder.py
