+++ parser/repository/qdrant/vectorizator.py
@@ -1,10 +1,11 @@
 import logging
 from typing import Optional, Dict, Any
 
-from infra.pdf_processor import PDFTextExtractor, TextChunker
-from infra.gemini_embeddings import GeminiEmbeddingService
-from infra.qdrant_client import QdrantVectorStore
-from domain.ports import VectorizationService
+from parser.gateway.pdf.pdf_processor import PDFTextExtractor, TextChunker
+from parser.gateway.gemini.embedder import GeminiEmbeddingService
+from parser.repository.qdrant.vector_store import QdrantVectorStore
+from parser.usecase.interfaces import VectorizationService
+
 
 logger = logging.getLogger(__name__)
