+++ parser/usecase/reports.py
@@ -1,25 +1,25 @@
 import logging
+
 from typing import cast
 
-from infra.e_disclosure import EDisclosureClient
-from infra.s3_storage import S3ReportsStorage
-from infra.db_repo import ReportsRepository
-from infra.models import ReportORM
 from companies import get_ticker_by_inn
-from domain.processing_result import ProcessingResult, ProcessingError, ReportMetadata, SingleCompanyProcessingResult
-from domain.report import DownloadedReport
-from parser.domain.ports import VectorizationService
+
+from parser.domain.model.processing_result import ProcessingResult, ProcessingError, ReportMetadata, SingleCompanyProcessingResult
+from parser.domain.model.report import DownloadedReport, ReportEntity
+
+from parser.usecase.interfaces import ReportsParser, VectorizationService, ReportsRepository, ColdStorage
 
 logger = logging.getLogger(__name__)
-
 
 class ReportProcessor:
     def __init__(
         self,
-        s3_client: S3ReportsStorage,
+        client: ReportsParser,
+        s3_client: ColdStorage,
         repo: ReportsRepository,
         vectorization_service: VectorizationService,
     ):
+        self.client = client
         self.s3_client = s3_client
         self.repo = repo
         self.vectorization_service = vectorization_service
@@ -31,22 +31,20 @@ class ReportProcessor:
             saved=0
         )
 
-        with EDisclosureClient() as client:
-            for inn in companies_inn:
-                try:
-                    result = self.process_company_by_inn(client, inn, skip_indexing=skip_indexing)
-                    results.processed += 1
-                    results.saved += result.saved
-                except Exception as e:
-                    logger.error(f"error while processing {inn}: {e}")
-                    results.errors.append(ProcessingError(inn=inn, error=str(e)))
+        for inn in companies_inn:
+            try:
+                result = self.process_company_by_inn(inn, skip_indexing=skip_indexing)
+                results.processed += 1
+                results.saved += result.saved
+            except Exception as e:
+                logger.error(f"error while processing {inn}: {e}")
+                results.errors.append(ProcessingError(inn=inn, error=str(e)))
 
         return results
 
-    def process_company_by_inn(self, client: EDisclosureClient, inn: str, skip_indexing: bool = False) -> SingleCompanyProcessingResult:
+    def process_company_by_inn(self, inn: str, skip_indexing: bool = False) -> SingleCompanyProcessingResult:
         ticker = get_ticker_by_inn(inn)
         return self.process_company_by_query(
-            client,
             query=inn,
             ticker=ticker,
             skip_indexing=skip_indexing,
@@ -54,12 +52,11 @@ class ReportProcessor:
 
     def process_company_by_query(
         self,
-        client: EDisclosureClient,
         query: str,
         ticker: str,
         skip_indexing: bool = False,
     ) -> SingleCompanyProcessingResult:
-        companies = client.search_company(query)
+        companies = self.client.search_company(query)
 
         if not companies:
             logger.warning(f"company not found: {query}")
@@ -69,7 +66,7 @@ class ReportProcessor:
         logger.info(f"selected company: {first_company['name']} (ID: {first_company['id']}, {ticker})")
 
         logger.info("downloading reports for %s...", ticker)
-        reports = client.download_reports(first_company)
+        reports = self.client.download_reports(first_company)
         downloaded = self._convert_to_downloaded_reports(reports)
 
         saved = self.process_downloaded_reports(downloaded, ticker, skip_indexing=skip_indexing)
@@ -113,35 +110,36 @@ class ReportProcessor:
             logger.error("error while saving to s3")
             return None
 
-        report_orm = self._save_report_to_db(ticker, report.year, report.period, s3_path)
+        report_orm = self._save_report_to_db(ticker, report.year, report.period, s3_path) # type: ignore
         if not report_orm:
             return None
 
         if not skip_indexing:
             self._vectorize_report(report_orm, report, ticker)
 
-        return ReportMetadata(s3_path=s3_path, year=report.year, period=report.period)
+        return ReportMetadata(s3_path=s3_path, year=report.year, period=report.period) # type: ignore
 
     def _upload_to_s3(self, ticker: str, report: DownloadedReport) -> str | None:
-        s3_path = self.s3_client.upload_report(ticker, report.year, report.period, report.path)
+        s3_path = self.s3_client.upload_report(ticker, report.year, report.period, report.path) # type: ignore
         if s3_path is None:
-            logger.error("Ошибка загрузки в S3")
+            logger.error("error uploading in s3")
             return None
-        logger.info(f"  Загружено в S3: {s3_path}")
+
+        logger.info(f"uploaded into S3: {s3_path}")
         return s3_path
 
-    def _vectorize_report(self, report_orm: ReportORM, report: DownloadedReport, ticker: str) -> None:
+    def _vectorize_report(self, report_orm: ReportEntity, report: DownloadedReport, ticker: str) -> None:
         result = self.vectorization_service.vectorize_report(
             report_id=cast(int, report_orm.id),
             file_path=report.path,
             ticker=ticker,
-            year=report.year,
+            year=report.year, # type: ignore
             period=report.period,
         )
         if not result.get("success"):
             logger.warning(f"Ошибка векторизации отчёта {ticker} {report.year}: {result.get('error')}")
 
-    def _save_report_to_db(self, ticker: str, year: int, period: str, s3_path: str) -> ReportORM | None:
+    def _save_report_to_db(self, ticker: str, year: int, period: str, s3_path: str) -> ReportEntity | None:
         try:
             report = self.repo.create_report(ticker, year, period, s3_path)
             if report is None:
@@ -158,7 +156,7 @@ class ReportProcessor:
         logger.info(f"{saved} reports saved")
 
     def _ensure_s3_uploaded(self, ticker, report) -> str | None:
-        existing_s3_path = self.s3_client.get_s3_report_link(ticker, report.year, report.period)
+        existing_s3_path = self.s3_client.get_report_link(ticker, report.year, report.period)
         if existing_s3_path:
             logger.info(f"skipping: report already exists in S3: {existing_s3_path}")
             return existing_s3_path
